Silence pytest numpy warnings with 3rd party libs
This silences a few warnings output when running
pytest tests caused by use of np.int etc.
in third party libraries, which is deprecated:

https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

It will be up to the libraries in question to fix,
so not relevant for us.

I also updated the deprecated --strict pytest
option to --strict-markers